Fix bug in range queries, finalize range and kNN plots.
Fix bug in range queries that led the program into a
never ending while loop  when the start value was
bigger that the stop value of the search.

Finalize the scripts that create the plots for
kNN and range queries.

Minor tweaks in Chord.py, FingerTable.py and Node.py

Signed-off-by: Jason Pavlakis <jasonpavlakis@gmail.com>